Move stepper stuff to stepper.rs
